menu de items con sus categorias

+++ src/app/api/menu-items/route.js
@@ -1,33 +1,39 @@
-import { MenuItem } from '@/models/MenuItem';
-import mongoose from 'mongoose';
+import { isAdmin } from "@/app/api/auth/[...nextauth]/route";
+import { MenuItem } from "@/models/MenuItem";
+import mongoose from "mongoose";
 
-export const POST = async (req) => {
+export async function POST(req) {
     mongoose.connect(process.env.MONGO_URL);
     const data = await req.json();
-    const menuItem = await MenuItem.create(data);
-    return Response.json(menuItem);
+    if (await isAdmin()) {
+        const menuItemDoc = await MenuItem.create(data);
+        return Response.json(menuItemDoc);
+    } else {
+        return Response.json({});
+    }
 }
 
-export const PUT = async (req) => {
+export async function PUT(req) {
     mongoose.connect(process.env.MONGO_URL);
-    const { _id, ...data } = await req.json();
-    await MenuItem.findByIdAndUpdate(_id, data);
+    if (await isAdmin()) {
+        const { _id, ...data } = await req.json();
+        await MenuItem.findByIdAndUpdate(_id, data);
+    }
     return Response.json(true);
-
 }
 
-
-export const GET = async (req) => {
+export async function GET() {
     mongoose.connect(process.env.MONGO_URL);
     return Response.json(
         await MenuItem.find()
-    )
+    );
 }
 
-export const DELETE = async (req) => {
+export async function DELETE(req) {
     mongoose.connect(process.env.MONGO_URL);
     const url = new URL(req.url);
     const _id = url.searchParams.get('_id');
-    await MenuItem.deleteOne({ _id });
-    return Response.json(true);
+    if (await isAdmin()) {
+        await MenuItem.deleteOne({ _id });
+    }
 }
